chore: 分类signals结构
